# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PedagangPulsa API
  description: |
    PedagangPulsa Member Mobile API v1.0

    ## Authentication

    The API uses JWT (JSON Web Token) based authentication with a 2-step verification process:

    1. **Login** → Get Access Token (15min) + Refresh Token (7days)
    2. **PIN Verify** → Get PIN Session Token (5min) for sensitive operations

    ## Idempotency

    Transaction and Transfer endpoints require `X-Reference-Id` header to prevent duplicate requests.

  version: 1.0.0
  contact:
    name: PedagangPulsa API Support
    email: api@pedagangpulsa.com
    url: https://pedagangpulsa.com
  license:
    name: Proprietary
servers:
  - url: https://api.pedagangpulsa.com/api/v1
    description: Production server
  - url: https://sandbox-api.pedagangpulsa.com/api/v1
    description: Sandbox server
...
paths:
...
  /transactions:
    post:
      tags:
        - Transactions
      summary: Create transaction
      description: |
        Purchase a product.
        Requires both X-Reference-Id header (for idempotency) AND pin_session_token in request body.
      operationId: createTransaction
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/X-Reference-Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
            example:
              product_id: 550e8400-e29b-41d4-a716-446655440005
              destination: "08123456789"
              pin_session_token: 550e8400-e29b-41d4-a716-446655440001
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Insufficient balance or invalid PIN session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Transactions
      summary: List transactions
      description: Get user's transaction history
      operationId: listTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, success, failed]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TransactionListResponse'
  /transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Get transaction detail
      description: Get detailed transaction information including attempt timeline
      operationId: getTransactionDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction detail retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TransactionDetail'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
  parameters:
    X-Reference-Id:
      name: X-Reference-Id
      in: header
      required: true
      description: |
        Unique reference ID for idempotency.
        Format: app-{user_id}-{timestamp}-{random}
        Example: app-550e8400-e29b-41d4-a716-446655440000-1710234567-a1b2c3
      schema:
        type: string
        example: app-550e8400-e29b-41d4-a716-446655440000-1710234567-a1b2c3
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - success
        - error_code
        - message
      properties:
        success:
          type: boolean
          example: false
        error_code:
          type: string
          example: INSUFFICIENT_BALANCE
        message:
          type: string
          example: Saldo tidak mencukupi
        details:
          type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8

    # Auth schemas
...
    CreateTransactionRequest:
      type: object
      required:
        - product_id
        - destination
        - pin_session_token
      properties:
        product_id:
          type: string
          format: uuid
        destination:
          type: string
          description: Destination number (phone, meter number, etc.)
        pin_session_token:
          type: string
          format: uuid
          description: Token from /auth/pin/verify
    TransactionResponse:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        reference_id:
          type: string
        status:
          type: string
          enum: [pending, processing, success, failed, refunded]
        product:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            operator:
              type: string
            type:
              type: string
        destination:
          type: string
        price:
          type: integer
        created_at:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    TransactionDetail:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        reference_id:
          type: string
        status:
          type: string
        product:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            sku:
              type: string
            operator:
              type: string
            category:
              type: string
            type:
              type: string
        destination:
          type: string
        price:
          type: integer
        cost_price:
          type: integer
        profit:
          type: integer
        sn:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        attempts:
          type: array
          items:
            type: object
            properties:
              attempt_id:
                type: string
                format: uuid
              supplier_name:
                type: string
              seq:
                type: integer
              status:
                type: string
              created_at:
                type: string
                format: date-time
              completed_at:
                type: string
                format: date-time
              supplier_trx_id:
                type: string
              sn:
                type: string
              error_code:
                type: string
              error_message:
                type: string

    # Transfer schemas
...